+++ sdk/ml/azure-ai-ml/azure/ai/ml/constants/_common.py
@@ -166,6 +166,7 @@ SPARK_ENVIRONMENT_WARNING_MESSAGE = (
     "Spark job will only install the packages defined in the Conda configuration. It "
     "will not create a docker container using the image defined in the environment."
 )
+NEW_ENTITY_PRINT_KWARG = "user_friendly_print_style"
 
 
 class AzureMLResourceType(object):